requirements.txt
python-telegram-bot==13.15
openai
pandas
openpyxl
gspread
oauth2client
python-dotenv
